lib/core: Make the start of the debug output more clear
I missed the quiet "Log start" between the test failure and the debug
output (and so was very confused by the repetition). You have to shout
at me!

v2: Thomas suggested I make the end of the logging clear as well. Since
I completely missed that there was a end marker, he must be right!

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Thomas Wood <thomas.wood@intel.com>