igt/gem_exec_lut_handle: Hide exec latency
The first test tries to rewrite relocations in an active batch, which is
a useful test and measurement. However, the overhead of the exec may
dominate and so we want a measurement without that overhead as well.
Using a pool of batches should allow for the oldest to idle whilst we
setup the next (and so the wait should be non-existent).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>