tests/kms_fbc_crc: exec_nop() can also invalidate FBC
So make sure that, at prepare_test(), we wait for FBC to be enabled
again after we run the exec_nop() call. Since after this happens, we
just assert fbc_enabled() at test_crc() instead of waiting for it to
be enabled.

This is now needed because we moved to software frontbuffer tracking,
so it can take some considerable time for FBC to be reenabled after it
is disabled.

A previous version of this patch was just calling gem_sync() after
exec_nop().

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>