intel_display_crc: A new tool to play with display CRCs
The CRC debug interface is a bit more than a simple textual file in
debugfs as there are a small command language to control what we want
from them.

This tool starts, slowly, by allowing us to dump the pipe CRCs whenever
we want. It can be handy to check what is the current CRC when we reach
a certain state on the screen (when using --interactive-debug for
instance) against a known CRC.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>