tests/pm_rc6_residency: remove redundant idle loops
Currently the test runs a separate idle loop when reading out each RC6
counter. But there is no need for this, we can have a signle idle loop
and read out all the counters at once.

This prepares for an upcoming patch where we need to consider the RC6P
and RC6PP counters as well when checking RC6.

Signed-off-by: Imre Deak <imre.deak@intel.com>
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>