kms_frontbuffer_tracking: reduce the FBC wait timeout to 2s
Just like we did for PSR, let's do it for FBC. FBC gets reenabled in
just 50ms, so the 5000ms timeout is huge. On the other hand, we only
pay the 5000ms timeout full price 9 times when running
kms_frontbuffer_tracking --fbc-only, so this change shouldn't save too
much time.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>